requirements.txt
Pillow==10.3.0
streamlit==1.35.0
torch==2.3.1
transformers==4.41.2
